Move SliceUtils.kt contents to SlicePermissions.kt
